a bit more circular dependency cleaning
